Move more of the verification code to verification.rs
