+++ core/consensus/babe/src/lib.rs
@@ -439,98 +439,6 @@ fn find_next_epoch_digest<B: BlockT>(header: &B::Header) -> Result<Option<Epoch>
 	Ok(epoch_digest)
 }
 
-/// Check a primary slot proposal header. We validate that the given header is
-/// properly signed by the expected authority, and that the contained VRF proof
-/// is valid. Additionally, the weight of this block must increase compared to
-/// its parent since it is a primary block.
-fn check_primary_header<H: Header>(
-	pre_hash: H::Hash,
-	pre_digest: (&VRFOutput, &VRFProof, AuthorityIndex, SlotNumber, BabeBlockWeight),
-	signature: AuthoritySignature,
-	parent_weight: BabeBlockWeight,
-	authorities: &[(AuthorityId, BabeAuthorityWeight)],
-	randomness: [u8; 32],
-	epoch_index: u64,
-	c: (u64, u64),
-) -> Result<(), String> {
-	let (vrf_output, vrf_proof, authority_index, slot_number, weight) = pre_digest;
-	if weight != parent_weight + 1 {
-		return Err("Invalid weight: should increase with Primary block.".into());
-	}
-
-	let author = &authorities[authority_index as usize].0;
-
-	if AuthorityPair::verify(&signature, pre_hash, &author) {
-		let (inout, _) = {
-			let transcript = make_transcript(
-				&randomness,
-				slot_number,
-				epoch_index,
-			);
-
-			schnorrkel::PublicKey::from_bytes(author.as_slice()).and_then(|p| {
-				p.vrf_verify(transcript, vrf_output, vrf_proof)
-			}).map_err(|s| {
-				babe_err!("VRF verification failed: {:?}", s)
-			})?
-		};
-
-		let threshold = calculate_primary_threshold(c, authorities, authority_index as usize);
-		if !check_primary_threshold(&inout, threshold) {
-			return Err(babe_err!("VRF verification of block by author {:?} failed: \
-								  threshold {} exceeded", author, threshold));
-		}
-
-		Ok(())
-	} else {
-		Err(babe_err!("Bad signature on {:?}", pre_hash))
-	}
-}
-
-/// Check a secondary slot proposal header. We validate that the given header is
-/// properly signed by the expected authority, which we have a deterministic way
-/// of computing. Additionally, the weight of this block must stay the same
-/// compared to its parent since it is a secondary block.
-fn check_secondary_header<H: Header>(
-	pre_hash: H::Hash,
-	pre_digest: (AuthorityIndex, SlotNumber, BabeBlockWeight),
-	signature: AuthoritySignature,
-	parent_weight: BabeBlockWeight,
-	authorities: &[(AuthorityId, BabeAuthorityWeight)],
-	randomness: [u8; 32],
-) -> Result<(), String> {
-	let (authority_index, slot_number, weight) = pre_digest;
-
-	if weight != parent_weight {
-		return Err("Invalid weight: Should stay the same with secondary block.".into());
-	}
-
-	// check the signature is valid under the expected authority and
-	// chain state.
-	let expected_author = secondary_slot_author(
-		slot_number,
-		authorities,
-		randomness,
-	).ok_or_else(|| "No secondary author expected.".to_string())?;
-
-	let author = &authorities[authority_index as usize].0;
-
-	if expected_author != author {
-		let msg = format!("Invalid author: Expected secondary author: {:?}, got: {:?}.",
-			expected_author,
-			author,
-		);
-
-		return Err(msg);
-	}
-
-	if AuthorityPair::verify(&signature, pre_hash.as_ref(), author) {
-		Ok(())
-	} else {
-		Err(format!("Bad signature on {:?}", pre_hash))
-	}
-}
-
 /// State that must be shared between the import queue and the authoring logic.
 #[derive(Default, Clone, Debug)]
 pub struct BabeLink(Arc<Mutex<(Option<Duration>, Vec<(Instant, u64)>)>>);
